pyproject.toml
[tool.poetry]
name = "webrock"
version = "0.0.1"
description = "AWS Bedrock + Playwright"
authors = ["Michael Scully"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.45.1"
pillow = "^10.4.0"
boto3 = "^1.34.153"
black = "^24.8.0"
ruff = "^0.5.6"
mypy = "^1.11.1"
langchain-core = "^0.2.28"
pydantic = "^2.8.2"
mypy-boto3-bedrock = "^1.34.152"
mypy-boto3-bedrock-runtime = "^1.34.149"
devtools = "^0.12.2"
mypy-boto3-s3 = "^1.35.2"
jupyter = "^1.1.1"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["webrock", "tests"]
known_first_party = ["commons"]

[tool.mypy]
strict = true
ignore_missing_imports = true
pretty = true
show_error_codes = true
implicit_reexport = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true
files = "webrock"
exclude = ["tests", ".venv"]

[tool.ruff]
# Same as Black.
line-length = 120
lint.ignore = [
    "E501",  # line too long, handled by black
    "C901",  # too complex
]

[tool.poetry.scripts]
webrock = "webrock:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
